Changed the way courses are displayed
Instead of a function creating a huge string of HTML code, the courses page now uses EJS's control flow functionality.

A JSON file of courses are sent to the EJS template for rendering, with an array of available sections for each course obtained before rendering.

Note that this approach feels a bit slow, but adding async functions into EJS was a bit complicated.

Also, note that lines 89-91 did not make sense, as it would render the page for only one course. It wasn't visible during testing since at this point, functionality is limited to one course per search.