Added favorites functionality
Allows users to favorite certain course sections and display them all in a singular page. All favorite courses are stored in MongoDB.

There are a few things to add, such as favorite classes will not show up as favorited in course searches, which will result in the same course possibly being favorited twice. There must be a control flow that checks for this whenever a search is made.

Also, at the time of writing, MongoDB takes very long to return a response, making the favorites button light up after quite a while. Need faster alternatives.