Streamlined Results Displaying Process
To prevent long wait times when a user searches for courses (especially in large bundles of them), the retrieval of sections information is only executed when the user demands to see it. This happens when the user clicks on the 'Sections' accordion button within the page.

Also added a cute little spinning circle to show the user the testudo is working hard.