Introduced Sorting and Filters
Users can now sort and filter out the courses that they see on the results page. Sorting options include course IDs and credits, whereas filter options include credits and gen-ed credit types.

All the logic to be used in this page will now be in the courseBuilder.js script, as there will be more future implementations to be added (so excited to do this).

New styles were introduced to format the new feature, and some added some underline animation to the navbar elements.

Might also consider a minimize feature to help reduce the number of displayed content for each course. It could help users, for example, discern which courses they might not be interested in.